# clients/typescript/openapi.json
{
  "components": {
    "schemas": {
      "ActionReason": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:/@-]+$",
            "title": "Code",
            "type": "string"
          },
          "message": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ActionReason",
        "type": "object"
      },
      "ActionRequest": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "discriminator": {
              "mapping": {
                "filesystem": "#/components/schemas/FilesystemAction",
                "network": "#/components/schemas/NetworkAction",
                "process": "#/components/schemas/ProcessAction"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/ProcessAction"
              },
              {
                "$ref": "#/components/schemas/FilesystemAction"
              },
              {
                "$ref": "#/components/schemas/NetworkAction"
              }
            ],
            "title": "Action"
          },
          "format": {
            "const": "polaris.action-review/0.1.0",
            "default": "polaris.action-review/0.1.0",
            "title": "Format",
            "type": "string"
          }
        },
        "required": [
          "action"
        ],
        "title": "ActionRequest",
        "type": "object"
      },
      "ActionReview": {
        "additionalProperties": false,
        "properties": {
          "action_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Action Digest",
            "type": "string"
          },
          "authorized": {
            "const": false,
            "default": false,
            "title": "Authorized",
            "type": "boolean"
          },
          "executed": {
            "const": false,
            "default": false,
            "title": "Executed",
            "type": "boolean"
          },
          "format": {
            "const": "polaris.action-review/0.1.0",
            "default": "polaris.action-review/0.1.0",
            "title": "Format",
            "type": "string"
          },
          "policy_changed": {
            "const": false,
            "default": false,
            "title": "Policy Changed",
            "type": "boolean"
          },
          "policy_digest": {
            "anyOf": [
              {
                "pattern": "^sha256:[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Digest"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/ActionReason"
            },
            "maxItems": 16,
            "minItems": 1,
            "title": "Reasons",
            "type": "array"
          },
          "risk": {
            "enum": [
              "low",
              "elevated",
              "unknown"
            ],
            "title": "Risk",
            "type": "string"
          },
          "safer_alternative": {
            "anyOf": [
              {
                "maxLength": 4096,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Safer Alternative"
          },
          "scope_only": {
            "const": true,
            "default": true,
            "title": "Scope Only",
            "type": "boolean"
          },
          "status": {
            "enum": [
              "within_declared_scope",
              "out_of_scope",
              "needs_review"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "action_digest",
          "status",
          "risk",
          "reasons"
        ],
        "title": "ActionReview",
        "type": "object"
      },
...
      "ApiError": {
        "additionalProperties": false,
        "description": "Every error that isn't an assessment-contract envelope. Never echoes submitted code.",
        "properties": {
          "code": {
            "enum": [
              "unauthorized",
              "invalid_host",
              "not_found",
              "method_not_allowed",
              "unsupported_media_type",
              "payload_too_large",
              "too_many_files",
              "invalid_json",
              "invalid_request",
              "rate_limited",
              "queue_full",
              "model_unavailable",
              "internal_error"
            ],
            "title": "Code",
            "type": "string"
          },
          "fields": {
            "description": "Invalid field names, never values.",
            "items": {
              "type": "string"
            },
            "title": "Fields",
            "type": "array"
          },
          "kind": {
            "const": "api_error",
            "default": "api_error",
            "title": "Kind",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "CandidateRequest": {
        "additionalProperties": false,
        "properties": {
          "edits": {
            "items": {
              "$ref": "#/components/schemas/CandidateEdit"
            },
            "maxItems": 8,
            "minItems": 1,
            "title": "Edits",
            "type": "array"
          },
          "expected_snapshot_digest": {
            "anyOf": [
              {
                "pattern": "^sha256:[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Snapshot Digest"
          },
          "rationale": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Rationale",
            "type": "string"
          },
          "verification_commands": {
            "items": {
              "$ref": "#/components/schemas/ProcessAction"
            },
            "maxItems": 16,
            "title": "Verification Commands",
            "type": "array"
          }
        },
        "required": [
          "edits",
          "rationale"
        ],
        "title": "CandidateRequest",
        "type": "object"
      },
      "CapabilityManifest": {
        "additionalProperties": false,
        "properties": {
          "analyzers": {
            "items": {
              "$ref": "#/components/schemas/AnalyzerCapability"
            },
            "title": "Analyzers",
            "type": "array"
          },
          "default_checks": {
            "items": {
              "type": "string"
            },
            "title": "Default Checks",
            "type": "array"
          },
          "format": {
            "const": "polaris.capabilities/0.2.0",
            "default": "polaris.capabilities/0.2.0",
            "title": "Format",
            "type": "string"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "matrix": {
            "items": {
              "$ref": "#/components/schemas/CheckCapability"
            },
            "title": "Matrix",
            "type": "array"
          },
          "workflow_format": {
            "const": "polaris.review/0.2.0",
            "default": "polaris.review/0.2.0",
            "title": "Workflow Format",
            "type": "string"
          }
        },
        "required": [
          "default_checks",
          "analyzers",
          "matrix",
          "limitations"
        ],
        "title": "CapabilityManifest",
        "type": "object"
      },
...
      "FilesystemAction": {
        "additionalProperties": false,
        "properties": {
          "action_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:/@-]+$",
            "title": "Action Id",
            "type": "string"
          },
          "kind": {
            "const": "filesystem",
            "default": "filesystem",
            "title": "Kind",
            "type": "string"
          },
          "operation": {
            "enum": [
              "read",
              "write",
              "delete"
            ],
            "title": "Operation",
            "type": "string"
          },
          "path": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "action_id",
          "operation",
          "path"
        ],
        "title": "FilesystemAction",
        "type": "object"
      },
...
      "NetworkAction": {
        "additionalProperties": false,
        "properties": {
          "action_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:/@-]+$",
            "title": "Action Id",
            "type": "string"
          },
          "kind": {
            "const": "network",
            "default": "network",
            "title": "Kind",
            "type": "string"
          },
          "method": {
            "enum": [
              "GET",
              "HEAD",
              "POST",
              "PUT",
              "PATCH",
              "DELETE"
            ],
            "title": "Method",
            "type": "string"
          },
          "url": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "action_id",
          "method",
          "url"
        ],
        "title": "NetworkAction",
        "type": "object"
      },
      "PatchProposal": {
        "additionalProperties": false,
        "properties": {
          "changed_lines": {
            "maximum": 2000.0,
            "minimum": 1.0,
            "title": "Changed Lines",
            "type": "integer"
          },
          "diff": {
            "maxLength": 262144,
            "minLength": 1,
            "title": "Diff",
            "type": "string"
          },
          "edits": {
            "items": {
              "$ref": "#/components/schemas/CandidateEdit"
            },
            "maxItems": 16,
            "minItems": 1,
            "title": "Edits",
            "type": "array"
          },
          "format": {
            "const": "polaris.proposal/0.1.0",
            "default": "polaris.proposal/0.1.0",
            "title": "Format",
            "type": "string"
          },
          "origin": {
            "default": "host_candidate",
            "enum": [
              "host_candidate",
              "configured_generator"
            ],
            "title": "Origin",
            "type": "string"
          },
          "proposal_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Proposal Digest",
            "type": "string"
          },
          "rationale": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Rationale",
            "type": "string"
          },
          "snapshot": {
            "$ref": "#/components/schemas/ReviewedSnapshot"
          },
          "verification_commands": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ProcessAction"
            },
            "maxItems": 16,
            "title": "Verification Commands",
            "type": "array"
          }
        },
        "required": [
          "snapshot",
          "edits",
          "rationale",
          "diff",
          "changed_lines",
          "proposal_digest"
        ],
        "title": "PatchProposal",
        "type": "object"
      },
      "ProcessAction": {
        "additionalProperties": false,
        "properties": {
          "action_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:/@-]+$",
            "title": "Action Id",
            "type": "string"
          },
          "argv": {
            "default": [],
            "items": {
              "maxLength": 4096,
              "pattern": "^[^\\x00\\r\\n]*$",
              "type": "string"
            },
            "maxItems": 128,
            "title": "Argv",
            "type": "array"
          },
          "cwd": {
            "default": ".",
            "maxLength": 512,
            "minLength": 1,
            "title": "Cwd",
            "type": "string"
          },
          "executable": {
            "maxLength": 512,
            "minLength": 1,
            "pattern": "^[^\\s\\x00;&|`$<>]+$",
            "title": "Executable",
            "type": "string"
          },
          "filesystem_targets": {
            "default": [],
            "items": {
              "maxLength": 512,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 64,
            "title": "Filesystem Targets",
            "type": "array"
          },
          "kind": {
            "const": "process",
            "default": "process",
            "title": "Kind",
            "type": "string"
          },
          "network_targets": {
            "default": [],
            "items": {
              "maxLength": 4096,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 32,
            "title": "Network Targets",
            "type": "array"
          }
        },
        "required": [
          "action_id",
          "executable"
        ],
        "title": "ProcessAction",
        "type": "object"
      },
...
      "WorkflowEnvelope": {
        "additionalProperties": false,
        "properties": {
          "changes": {
            "items": {
              "$ref": "#/components/schemas/ChangeRecord"
            },
            "title": "Changes",
            "type": "array"
          },
          "context": {
            "$ref": "#/components/schemas/ContextSummary"
          },
          "finding_count": {
            "title": "Finding Count",
            "type": "integer"
          },
          "format": {
            "const": "polaris.workflow/0.1.0",
            "default": "polaris.workflow/0.1.0",
            "title": "Format",
            "type": "string"
          },
          "notices": {
            "items": {
              "type": "string"
            },
            "title": "Notices",
            "type": "array"
          },
          "report_id": {
            "title": "Report Id",
            "type": "string"
          },
          "review": {
            "$ref": "#/components/schemas/WorkflowReviewReport"
          },
          "snapshot": {
            "$ref": "#/components/schemas/SnapshotReference"
          },
          "status": {
            "enum": [
              "complete",
              "incomplete",
              "stale",
              "error"
            ],
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "tests_status": {
            "const": "not_run",
            "default": "not_run",
            "title": "Tests Status",
            "type": "string"
          }
        },
        "required": [
          "report_id",
          "status",
          "summary",
          "finding_count",
          "snapshot",
          "changes",
          "context",
          "review"
        ],
        "title": "WorkflowEnvelope",
        "type": "object"
      },
...
      "WorkflowRepairRequest": {
        "additionalProperties": false,
        "properties": {
          "candidate": {
            "$ref": "#/components/schemas/CandidateRequest"
          },
          "config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkflowReviewConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/WorkflowFile"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Files",
            "type": "array"
          }
        },
        "required": [
          "files",
          "candidate"
        ],
        "title": "WorkflowRepairRequest",
        "type": "object"
      },
...
      "WorkflowReviewRequest": {
        "additionalProperties": false,
        "properties": {
          "config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkflowReviewConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/WorkflowFile"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Files",
            "type": "array"
          }
        },
        "required": [
          "files"
        ],
        "title": "WorkflowReviewRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "An API key from `polaris serve keys create`. Required when the server was started with keys.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Versioned security review and bounded engineering proposals. Legacy Python/model contracts remain separate.\n\n* `POST /v1/review` takes a diff, whole files or a snippet and returns findings (JSON or SARIF).\n* `POST /v1/assess` runs the `polaris.assessment/0.1.0` contract; `POST /v1/assess/batch` runs\n  up to 64 of them at once (the hosted-model client uses it).\n* Findings estimate risk; they never approve, block or authorize anything.\n* `/v1/workflow/review` adds explicit coverage and content binding; `/v1/workflow/propose`\n  validates host candidates. `/v1/workflow/action` assesses configured scope without execution.\n* Project code is parsed, never run. The workflow API defaults to memory-only analysis.\n  An administrator may explicitly enable isolated transient source files for external analyzers.\n  Request bodies are not logged or persistently stored; proposal responses contain scoped edits.\n  No remote apply or arbitrary shell endpoint exists.\n\nSend `Authorization: Bearer <key>` when the server was started with API keys.\n",
    "summary": "Polaris by TheoVex: security review and bounded engineering proposals.",
    "title": "Polaris API",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/workflow/action": {
      "post": {
        "operationId": "action_v1_workflow_action_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionReview"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid JSON or host."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Missing or invalid API key."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request too large."
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not JSON."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "The JSON doesn't match this endpoint."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Rate limited or server busy; see Retry-After."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "No model is loaded (engine \"model\")."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Assess an action using application-owned policy; no execution",
        "tags": [
          "workflow"
        ]
      }
    },
    "/v1/workflow/capabilities": {
      "get": {
        "operationId": "capabilities_v1_workflow_capabilities_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilityManifest"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid JSON or host."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Missing or invalid API key."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request too large."
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not JSON."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "The JSON doesn't match this endpoint."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Rate limited or server busy; see Retry-After."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "No model is loaded (engine \"model\")."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Actual workflow analyzer availability and limits",
        "tags": [
          "workflow"
        ]
      }
    },
    "/v1/workflow/propose": {
      "post": {
        "operationId": "propose_v1_workflow_propose_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowRepairRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchProposal"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid JSON or host."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Missing or invalid API key."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request too large."
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not JSON."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "The JSON doesn't match this endpoint."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Rate limited or server busy; see Retry-After."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "No model is loaded (engine \"model\")."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Validate supplied candidate edits; never apply or execute",
        "tags": [
          "workflow"
        ]
      }
    },
    "/v1/workflow/review": {
      "post": {
        "operationId": "review_v1_workflow_review_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid JSON or host."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Missing or invalid API key."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request too large."
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not JSON."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "The JSON doesn't match this endpoint."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Rate limited or server busy; see Retry-After."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "No model is loaded (engine \"model\")."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Review submitted content; no server paths or project execution",
        "tags": [
          "workflow"
        ]
      }
    }
  }
}